+++ mayan/apps/document_states/dependencies.py
@@ -1,5 +1,5 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='graphviz', version_string='==0.13.2'
+    module=__name__, name='graphviz', version_string='==0.14'
 )